chore: refactor file structure and function naming in cli package
- Rename the file `adapter/cli/app/version.go` to `adapter/cli/app/version/version.go`
- Change the function name `newVersionCmd` to `NewVersionCmd` in the file `adapter/cli/app/version/version.go`
- Remove the line `rootCmd.AddCommand(newVersionCmd())` and add `rootCmd.AddCommand(version.NewVersionCmd())` in the file `adapter/cli/app/cmd.go`
